requirements.txt
lxml
beautifulsoup4
jupyter
pandas
numpy
requests
cssselect
kafka-python
schedule
pyhocon
diskcache
redis